Implemented Spring Security, JWT Authentication, Users, Sign-in, Sign-up
Added User, Roles entities along with the services/dao,
Built an API that registers new users,
Built an API that allows users to log in,
The login API will generate a JWT authentication token that is returned in the response,
Configured Spring Security to restrict access to protected resources,
Configured Spring Security to throw a 401 error if client does not have a valid JWT token and tries to access
a protected resource,
Configured role based authentication.

Fetch calls were implemented on the client to test the API's, needs refactoring.